fix DbContext.cs
