--- ToDoList-WebApi.Server/db/DbContextUser.cs
@@ -1,21 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
- 
-namespace SQLite
-{
-    public class ApplicationContextUser : DbContext
-    {
-        public DbSet<SQLite.User> Users {get;set; } = null!;
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            optionsBuilder.UseSqlite("Data Source=./db/users.sqlite");
-        }
-        public User FindUserByName(string name)
-        {
-            return Users.FirstOrDefault(u => u.Name == name);
-        }
-        public User FindUserByNameAndPassword(string name, string password)
-        {
-            return Users.FirstOrDefault(u => u.Name == name && u.Password == password);
-        }
-    }
-}
